refactor(models): split models.py by domain into a package
models.py grew to 553 LOC with 24 classes spanning every part of the
API surface. Convert it to a models/ package with one file per domain
(alarm, installation, auth, lock, camera, activity, sentinel, services)
and re-export everything from models/__init__.py so every existing
import path keeps working unchanged. Pure file-level reorganisation —
no behaviour change. All tests pass.
