+++ custom_components/verisure_owa/verisure_owa_api/models/auth.py
@@ -1,0 +1,12 @@
+"""Auth-domain models."""
+
+from __future__ import annotations
+
+from pydantic import BaseModel
+
+
+class OtpPhone(BaseModel):
+    """OTP phone item for two-factor authentication."""
+
+    id: int
+    phone: str
